refactor: isolate namespace file client
